refactor(TossPaymentClientConfig): 클래스명 개선
